requirements.txt
numpy~=1.24.3
pillow~=10.2.0
torchvision~=0.17.0


tqdm~=4.65.0